TST/REF: collect tests for get_numeric_data (#37634)
* misplaced loc test

* TST/REF: collect get_numeric_data tests
