+++ pandas/tests/generic/test_frame.py
@@ -61,14 +61,6 @@ class TestDataFrame(Generic):
         with pytest.raises(ValueError, match=msg):
             bool(df)
 
-    def test_get_numeric_data_preserve_dtype(self):
-
-        # get the numeric data
-        o = DataFrame({"A": [1, "2", 3.0]})
-        result = o._get_numeric_data()
-        expected = DataFrame(index=[0, 1, 2], dtype=object)
-        self._compare(result, expected)
-
     def test_metadata_propagation_indiv_groupby(self):
         # groupby
         df = DataFrame(
